[HOTSPOT-92] fix: 스크롤 수정 및 파일 구조 변경
