+++ apps/web/src/app/(shell)/(main)/page.tsx
