+++ apps/web/src/app/(shell)/layout.tsx
@@ -1,0 +1,14 @@
+import { BottomNavController } from '@/widgets/app-bottom-nav';
+
+export default function MobileLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <div className="flex flex-col h-full overflow-auto pb-14.5">{children}</div>
+      <BottomNavController />
+    </>
+  );
+}
