+++ apps/web/src/widgets/app-bottom-nav/index.ts
@@ -1,0 +1,1 @@
+export { BottomNavController } from './ui/BottomNavController';
